Revert "Merge branch 'production' of https://github.com/specify/specify7 into production"
This reverts commit bd600dc0727b4ffb1ff8517edf2ffefc22e2d9af, reversing
changes made to 4cf4ee7882893b4cf3b737d685ce114efdaf1748.
