+++ specifyweb/frontend/js_src/lib/components/Forms/DeleteButton.tsx
@@ -44,7 +44,6 @@ export function DeleteButton<SCHEMA extends AnySchema>({
   deferred: initialDeferred = false,
   component: ButtonComponent = Button.Secondary,
   onDeleted: handleDeleted,
-  isIcon = false,
 }: {
   readonly resource: SpecifyResource<SCHEMA>;
   readonly deletionMessage?: React.ReactNode;
@@ -56,12 +55,10 @@ export function DeleteButton<SCHEMA extends AnySchema>({
   readonly deferred?: boolean;
   readonly component?: typeof Button['Secondary'];
   readonly onDeleted?: () => void;
-  readonly isIcon?: boolean;
 }): JSX.Element {
   const [deferred, setDeferred] = useLiveState<boolean>(
     React.useCallback(() => initialDeferred, [initialDeferred, resource])
   );
-
   const [blockers, setBlockers] = useAsyncState<RA<DeleteBlocker>>(
     React.useCallback(
       async () => (deferred ? undefined : fetchBlockers(resource)),
@@ -79,27 +76,16 @@ export function DeleteButton<SCHEMA extends AnySchema>({
 
   return (
     <>
-      {isIcon ? (
-        <Button.Icon
-          icon="trash"
-          title={isBlocked ? formsText.deleteBlocked() : commonText.delete()}
-          onClick={(): void => {
-            handleOpen();
-            setDeferred(false);
-          }}
-        />
-      ) : (
-        <ButtonComponent
-          title={isBlocked ? formsText.deleteBlocked() : undefined}
-          onClick={(): void => {
-            handleOpen();
-            setDeferred(false);
-          }}
-        >
-          {isBlocked ? icons.exclamation : undefined}
-          {commonText.delete()}
-        </ButtonComponent>
-      )}
+      <ButtonComponent
+        title={isBlocked ? formsText.deleteBlocked() : undefined}
+        onClick={(): void => {
+          handleOpen();
+          setDeferred(false);
+        }}
+      >
+        {isBlocked ? icons.exclamation : undefined}
+        {commonText.delete()}
+      </ButtonComponent>
       {isOpen ? (
         blockers === undefined ? (
           <Dialog
