+++ specifyweb/frontend/js_src/lib/components/TreeView/Row.tsx
@@ -7,10 +7,10 @@ import type { RA } from '../../utils/types';
 import { Button } from '../Atoms/Button';
 import { className } from '../Atoms/className';
 import { icons } from '../Atoms/Icons';
-import { getPref } from '../InitialContext/remotePrefs';
 import { userPreferences } from '../Preferences/userPreferences';
 import type { Conformations, KeyAction, Row, Stats } from './helpers';
 import { formatTreeStats, mapKey, scrollIntoView } from './helpers';
+import { getPref } from '../InitialContext/remotePrefs';
 
 export function TreeRow({
   row,
@@ -175,9 +175,9 @@ export function TreeRow({
                 color:
                   typeof row.acceptedId === 'number' ? synonymColor : undefined,
               }}
-              onClick={({ metaKey, shiftKey }): void => {
-                metaKey || shiftKey ? handleFocusNode([]) : handleToggle(false);
-              }}
+              onClick={({ metaKey, shiftKey }): void =>
+                metaKey || shiftKey ? handleFocusNode([]) : handleToggle(false)
+              }
               onKeyDown={(event): void => {
                 const action = mapKey(event);
                 if (action === undefined) return undefined;
