+++ specifyweb/frontend/js_src/lib/utils/cache/definitions.ts
@@ -75,9 +75,6 @@ export type CacheDefinitions = {
   } & {
     readonly /** Open nodes in a given tree */
     [key in `conformations${AnyTree['tableName']}`]: Conformations;
-  } & {
-    readonly isSplit: boolean;
-    readonly isHorizontal: boolean;
   };
   readonly workBenchSortConfig: {
     readonly /**
