+++ specifyweb/frontend/js_src/package.json
@@ -51,7 +51,6 @@
     "leaflet-gesture-handling": "^1.2.2",
     "leaflet.featuregroup.subgroup": "^1.0.2",
     "leaflet.markercluster": "^1.5.3",
-    "m-react-splitters": "^1.2.0",
     "rc-slider": "^10.1.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
